+++ src/main/java/ee/ut/cs/dsg/StreamCardinality/ApproximateAggregateFunction/HyperLogLogAggregationFunction.java
@@ -6,7 +6,7 @@ import org.apache.flink.api.common.functions.AggregateFunction;
 import org.apache.flink.api.java.tuple.Tuple3;
 import scala.collection.generic.BitOperations;
 
-public class HyperLogLogAggregationFunction implements AggregateFunction<Tuple3<Long, String, Integer>, HyperLogLogAccumulator, Tuple3<Long,String,Integer>> {
+public class HyperLogLogAggregationFunction implements AggregateFunction<Tuple3<Long, String, Long>, HyperLogLogAccumulator, Tuple3<Long,String,Long>> {
     public HyperLogLogAccumulator createAccumulator() { return new HyperLogLogAccumulator(); }
 
     public HyperLogLogAccumulator merge(HyperLogLogAccumulator a, HyperLogLogAccumulator b) {
@@ -18,7 +18,7 @@ public class HyperLogLogAggregationFunction implements AggregateFunction<Tuple3<
         return a;
     }
 
-    public HyperLogLogAccumulator add(Tuple3<Long, String, Integer> value, HyperLogLogAccumulator acc) {
+    public HyperLogLogAccumulator add(Tuple3<Long, String, Long> value, HyperLogLogAccumulator acc) {
         acc.f0 = value.f0;
         acc.f1 = value.f1;
         long val = Math.round(value.f2);
@@ -27,11 +27,11 @@ public class HyperLogLogAggregationFunction implements AggregateFunction<Tuple3<
         return acc;
     }
 
-    public Tuple3<Long, String, Integer> getResult(HyperLogLogAccumulator acc) {
-        Tuple3<Long,String,Integer> res= new Tuple3<>();
+    public Tuple3<Long, String, Long> getResult(HyperLogLogAccumulator acc) {
+        Tuple3<Long,String,Long> res= new Tuple3<>();
         res.f0 = acc.f0;
         res.f1 = acc.f1;
-        res.f2 = (int) acc.acc.cardinality();
+        res.f2 = acc.acc.cardinality();
 
         return res;
     }
