+++ src/main/java/ee/ut/cs/dsg/StreamCardinality/ApproximateAggregateFunction/HyperLogLogPlusAggregationFunction.java
@@ -5,7 +5,7 @@ import ee.ut.cs.dsg.StreamCardinality.ApproximateCardinality.HyperLogLogPlus;
 import org.apache.flink.api.common.functions.AggregateFunction;
 import org.apache.flink.api.java.tuple.Tuple3;
 
-public class HyperLogLogPlusAggregationFunction implements AggregateFunction<Tuple3<Long, String, Integer>, HyperLogLogPlusAccumulator, Tuple3<Long,String,Integer>> {
+public class HyperLogLogPlusAggregationFunction implements AggregateFunction<Tuple3<Long, String, Long>, HyperLogLogPlusAccumulator, Tuple3<Long,String,Long>> {
     @Override
     public HyperLogLogPlusAccumulator createAccumulator() { return new HyperLogLogPlusAccumulator(); }
 
@@ -20,7 +20,7 @@ public class HyperLogLogPlusAggregationFunction implements AggregateFunction<Tup
     }
 
     @Override
-    public HyperLogLogPlusAccumulator add(Tuple3<Long, String, Integer> value, HyperLogLogPlusAccumulator acc) {
+    public HyperLogLogPlusAccumulator add(Tuple3<Long, String, Long> value, HyperLogLogPlusAccumulator acc) {
         acc.f0 = value.f0;
         acc.f1 = value.f1;
         long val = Math.round(value.f2);
@@ -29,11 +29,11 @@ public class HyperLogLogPlusAggregationFunction implements AggregateFunction<Tup
     }
 
     @Override
-    public Tuple3<Long, String, Integer> getResult(HyperLogLogPlusAccumulator acc) {
-        Tuple3<Long,String,Integer> res = new Tuple3<>();
+    public Tuple3<Long, String, Long> getResult(HyperLogLogPlusAccumulator acc) {
+        Tuple3<Long,String,Long> res = new Tuple3<>();
         res.f0 = acc.f0;
         res.f1 = acc.f1;
-        res.f2 = (int) acc.acc.cardinality();
+        res.f2 = acc.acc.cardinality();
         return res;
     }
 }
