+++ src/main/java/ee/ut/cs/dsg/StreamCardinality/ApproximateAggregateFunction/LogLogAggregationFunction.java
@@ -3,7 +3,7 @@ package ee.ut.cs.dsg.StreamCardinality.ApproximateAggregateFunction;
 import org.apache.flink.api.common.functions.AggregateFunction;
 import org.apache.flink.api.java.tuple.Tuple3;
 
-public class LogLogAggregationFunction implements AggregateFunction<Tuple3<Long, String, Integer>, LogLogAccumulator, Tuple3<Long, String, Integer>> {
+public class LogLogAggregationFunction implements AggregateFunction<Tuple3<Long, String, Long>, LogLogAccumulator, Tuple3<Long, String, Long>> {
 
     public LogLogAccumulator createAccumulator() { return new LogLogAccumulator();}
 
@@ -12,7 +12,7 @@ public class LogLogAggregationFunction implements AggregateFunction<Tuple3<Long,
         return a;
     }
 
-    public LogLogAccumulator add(Tuple3<Long, String, Integer> value, LogLogAccumulator acc) {
+    public LogLogAccumulator add(Tuple3<Long, String, Long> value, LogLogAccumulator acc) {
         acc.f0 = value.f0;
         acc.f1 = value.f1;
         long val = Math.round(value.f2);
@@ -21,12 +21,12 @@ public class LogLogAggregationFunction implements AggregateFunction<Tuple3<Long,
         return acc;
     }
 
-    public Tuple3<Long, String, Integer> getResult(LogLogAccumulator acc) {
-        Tuple3<Long,String,Integer> res= new Tuple3<>();
+    public Tuple3<Long, String, Long> getResult(LogLogAccumulator acc) {
+        Tuple3<Long,String,Long> res= new Tuple3<>();
         res.f0 = acc.f0;
         res.f1 = acc.f1;
         try{
-            res.f2 = (int) acc.acc.cardinality();
+            res.f2 = acc.acc.cardinality();
         } catch (Exception e) {
             e.printStackTrace();
         }
